Incorporate asset files in expo publish
Need a reference to them in a require statement - this provides it.
TODO: Need to figure out how to actually load the Firebase bundle (certainly this won't work as-is and will need to use some react native features to resolve the return value from require()); also need to figure out how to use static mp3/m4v files now that they're included in the package.